+++ js/designer_components/DesignerComponent.ts
@@ -5,6 +5,8 @@ interface Window {
     Sk: any;
 }
 
+const HASUNITS = /[a-zA-Z%]/g
+
 /**
  * To create a designer class that supports dynamically updating the design view
  * subclass from DesignerComponent
@@ -19,7 +21,7 @@ export class DesignerComponent {
     static links: string[] = [];
 
     /** If using an external CDN to include a script tag override the script attribute */
-    static script: string[] = [];
+    static scripts: string[] = [];
     private static loaded = false;
     private static loading = false;
 
@@ -237,6 +239,11 @@ export class DesignerComponent {
             return color.slice(4, color.length - 1);
         }
         return color;
+    }
+
+    public getCssLength(len?: string) {
+        len ??= "";
+        return ("" + len).match(HASUNITS) ? len : len + "px";
     }
 
     /** helper method for clearing a domnode that will be reinstantiated during an update */